Refactor: Replace viewModel with hiltViewModel in LoginScreen and ProfileScreen
This commit replaces the `viewModel()` function with `hiltViewModel()` for instantiating ViewModels in `LoginScreen.kt` and `ProfileScreen.kt`.

Additionally, the `LoginScreen` now uses a `navigateToProfile` lambda to handle navigation to the Profile screen instead of directly relying on the ViewModel's sign-in state. The `MainActivity` is updated to use Jetpack Navigation Component with `NavHost` to manage navigation between `LoginScreen` and `ProfileScreen`.