feat: Add ProgressDialog composable
This commit introduces a new composable function `ProgressDialog`.

This composable displays a dialog containing a `CircularProgressIndicator` and a text message (`progressMessage`) to indicate an ongoing process. It is intended to be used as a progress indicator, with a `TODO` comment noting future improvement for inline progress.

The dialog is presented within a `Card` and uses a `Row` to align the `CircularProgressIndicator` and the message text.