Refactor: EventFolderRepository provider in RepositoryModule
feat: Clear event folders from DB on logout

This commit modifies the `ProfileViewmodel` to ensure that all event folders stored in the local Room database are deleted when a user logs out.

The `logout` function in `ProfileViewmodel` now calls `eventFolderRepository.removeAllFolderFromDB()` to clear the event folder data, in addition to clearing user data and preferences.