fix: Close DropdownMenu after action
This commit ensures that the `DropdownMenu` in the `EventCard` composable closes after a "Rename" or "Delete" action is performed.

Previously, the menu would remain open after clicking on an item. This change updates the `onClick` lambdas for "Rename" and "Delete" `DropdownMenuItem`s to set `dropDownMenuOpen` to `false`, thus closing the menu.